requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
mysql-connector-python==8.1.0
Werkzeug==2.3.7
Markdown==3.4.1
gunicorn==21.2.0
